Use a generator for getting all orders
Previously we would recursively build a list of orders until there were
none left, but this leaves us open to running out of memory if we ask
for too large of a date range to grab orders for.

By using a generator we keep only a single batch of orders in memory at
a time.