Dockerfile: update to official python image
At some point, `apk add python` stopped also installing pip3, so
building the image with the current Dockerfile was broken. Instead,
let's use a more stable environment that sets up python and pip for us
and lets us call them as just `python` and `pip`. We choose Python 3.8
because that's the latest version that the `mws` package supports, and
to make sure that our version of `mws` works with 3.8, we update `mws`
to the latest stable version.

Additionally, move the Dockerfile around so that we're making better use
of caching.