main: throttling, more logging, misc errors
This is a big one. Sorry.

Change the throttling strategy to wait long enough between requests so
that we only use up our quota slowly. In the event that we exhaust the
quota, wait long enough to refill our pool substantially and then
resume. Because most of our time is now spent sleeping between
`ListOrderItem` requests, we shouldn't really need to wait between
`ListOrders` requests, and it's unlikely that we'll run into throttling
there.

Add logging of the happy path throughout so that it's easier to see
what's happening. When we encounter an error, don't assume that it's due
to rate limiting and instead print out the entire error (it gets printed
out as XML).

Handle the case that the list of orders contains exactly one order in
the same way that we handle a list of exactly one item (the library
parses the XML into a dict rather than a list containing a dict).

Exchange some maps for a simple loop over the orders generator; note
that since we're iterating orders, we fetch 100 orders, then fill in the
items for them, then fetch up to another 100 orders, and so on until we
run out of orders.

Use `isinstance()` instead of `type() ==` because `pylint` prefers the
former.

Rename some methods and partials to be more intuitive (to me).