Makefile
Add a rudimentary Makefile so that it's a bit easier to keep the style
and linting up to date. Because we're running the linter outside of the
container, it's missing all of the requirements that we install within
the container. I don't think it should be too difficult to install
`pylint` in the container and run it there, but I didn't want to do it
right now.

Pylint is still complaining about some missing docblocks in addition to
unused imports, but no major complaints.