lint: fix naming errors
For reasons I don't entirely understand, pylint sees some of the
declarations at the top of `__main__.py` (but not, e.g., `orders_api`)
as constants, and it wants them to be uppercase. I think that's an
improvement because it makes them stand out as environment arguments.

It also doesn't like `fn` as a variable name because it's too short.
I agree less on this one, but it's easier to choose a more specific name
than to configure pylint for the exception.