Support explicit start and end dates
I found myself in a situation where I needed to backfill a one-day hole
more than 10 days in the past, and I got tired of waiting for the
backfill on all of the days that I didn't care about. Now if you specify
an explicit START_DATE, it overrides DAYS_AGO. If you specify END_DATE,
then we honor that; otherwise, we assume END_DATE is now.