black: stop depending on rando image
We can just as well install it ourselves as part of the "dev" image.
While I was making this change, I realized that we can just mount and
lint the working directory directly, in which case we don't need to copy
our source files into the image.

Note that we had to do more work than expected to install black
ourselves. There's no Alpine wheel for a black dependency that needs
compilation, so we end up having to compile it ourselves. The
`--virtual` and `apk del` business lets us install a "virtual" package
for the black dependencies and then delete that package after installing
black to keep the image size down (since we don't actually need the
build environment).

See https://github.com/psf/black/issues/1106
See https://stackoverflow.com/questions/46221063/what-is-build-deps-for-apk-add-virtual-command/46222036