Add missing guard for "NextToken" handling
We'd copy-pasted the code arount this to handle additional lists of
results when fetching from AWS. But we weren't handling the case where
we get a `NextToken` value but the next batch of results is empty. This
seems to happen when we have a number of results which is an even
multiple of 100. In that case, we'll see a `NextToken` link, fetch it,
and then crash because there's no results on it. Instead of crashing,
let's do the exact same thing we did up above if our first query found
no results.